--- src/main/java/flab/commercemarket/cart/dto/CartDto.java
@@ -1,21 +1,0 @@
-package flab.commercemarket.cart.dto;
-
-import flab.commercemarket.cart.domain.Cart;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class CartDto {
-    private long userId;
-    private long productId;
-    private int quantity;
-
-    public Cart toCart() {
-        return Cart.builder()
-                .userId(userId)
-                .productId(productId)
-                .quantity(quantity)
-                .build();
-    }
-}
