+++ src/main/java/flab/commercemarket/common/exception/DataNotFoundException.java
@@ -1,0 +1,11 @@
+package flab.commercemarket.common.exception;
+
+public class DataNotFoundException extends RuntimeException {
+    public DataNotFoundException(String message) {
+        super(message);
+    }
+
+    public DataNotFoundException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
