+++ src/main/java/flab/commercemarket/common/exception/DuplicateDataException.java
@@ -1,0 +1,11 @@
+package flab.commercemarket.common.exception;
+
+public class DuplicateDataException extends RuntimeException {
+    public DuplicateDataException(String message) {
+        super(message);
+    }
+
+    public DuplicateDataException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
