+++ src/main/java/flab/commercemarket/common/exception/ForbiddenException.java
@@ -1,0 +1,11 @@
+package flab.commercemarket.common.exception;
+
+public class ForbiddenException extends RuntimeException {
+    public ForbiddenException(String message) {
+        super(message);
+    }
+
+    public ForbiddenException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
