+++ src/main/java/flab/commercemarket/common/helper/AuthorizationHelper.java
@@ -1,0 +1,17 @@
+package flab.commercemarket.common.helper;
+
+import flab.commercemarket.common.exception.ForbiddenException;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
+
+@Slf4j
+@Component
+public class AuthorizationHelper {
+
+    public void checkUserAuthorization(long cartUserId, long loginUserId) {
+        if (cartUserId != loginUserId) {
+            log.info("cartUserId = {}, loginUserId = {}", cartUserId, loginUserId);
+            throw new ForbiddenException("유저 권한정보가 일치하지 않음");
+        }
+    }
+}
