+++ src/main/java/flab/commercemarket/controller/cart/dto/CartResponseDto.java
@@ -1,0 +1,17 @@
+package flab.commercemarket.controller.cart.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class CartResponseDto {
+    private Long id;
+    private long userId;
+    private long productId;
+    private int quantity;
+}
