+++ src/main/java/flab/commercemarket/controller/product/dto/LikeDto.java
@@ -1,0 +1,10 @@
+package flab.commercemarket.controller.product.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class LikeDto {
+    long productId;
+}
