+++ src/main/java/flab/commercemarket/controller/user/dto/UserResponseDto.java
@@ -1,0 +1,17 @@
+package flab.commercemarket.controller.user.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class UserResponseDto {
+
+    private Long id;
+    private String username;
+    private String password;
+    private String name;
+    private String address;
+    private String phoneNumber;
+    private String email;
+}
