--- src/main/java/flab/commercemarket/exception/DataNotFoundException.java
@@ -1,11 +1,0 @@
-package flab.commercemarket.exception;
-
-public class DataNotFoundException extends RuntimeException {
-    public DataNotFoundException(String message) {
-        super(message);
-    }
-
-    public DataNotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
