--- src/main/java/flab/commercemarket/exception/DuplicateDataException.java
@@ -1,11 +1,0 @@
-package flab.commercemarket.exception;
-
-public class DuplicateDataException extends RuntimeException {
-    public DuplicateDataException(String message) {
-        super(message);
-    }
-
-    public DuplicateDataException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
