--- src/main/java/flab/commercemarket/exception/ForbiddenException.java
@@ -1,11 +1,0 @@
-package flab.commercemarket.exception;
-
-public class ForbiddenException extends RuntimeException {
-    public ForbiddenException(String message) {
-        super(message);
-    }
-
-    public ForbiddenException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
