--- src/main/java/flab/commercemarket/product/dto/LikeDto.java
@@ -1,10 +1,0 @@
-package flab.commercemarket.product.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class LikeDto {
-    long productId;
-}
