--- src/main/java/flab/commercemarket/user/dto/UserResponseDto.java
@@ -1,17 +1,0 @@
-package flab.commercemarket.user.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class UserResponseDto {
-
-    private Long id;
-    private String username;
-    private String password;
-    private String name;
-    private String address;
-    private String phoneNumber;
-    private String email;
-}
