+++ src/main/resources/mybatis/cart-mapper.xml
@@ -1,26 +1,27 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper  PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"  "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="flab.commercemarket.cart.mapper.CartMapper">
-    <resultMap id="cartResultMap" type="flab.commercemarket.cart.domain.Cart">
-        <id property="id" column="id" />
-        <result property="userId" column="user_id" />
-        <result property="productId" column="product_id" />
-        <result property="quantity" column="quantity" />
+<mapper namespace="flab.commercemarket.domain.cart.mapper.CartMapper">
+    <resultMap id="cartResultMap" type="flab.commercemarket.domain.cart.vo.Cart">
+        <id property="id" column="id"/>
+        <result property="userId" column="user_id"/>
+        <result property="productId" column="product_id"/>
+        <result property="quantity" column="quantity"/>
     </resultMap>
 
-    <insert id="createCart" parameterType="flab.commercemarket.cart.domain.Cart" useGeneratedKeys="true" keyProperty="id">
+    <insert id="createCart" parameterType="flab.commercemarket.domain.cart.vo.Cart" useGeneratedKeys="true"
+            keyProperty="id">
         INSERT INTO cart (user_id, product_id, quantity)
         VALUES (#{userId}, #{productId}, #{quantity})
     </insert>
 
-    <update id="updateCart" parameterType="flab.commercemarket.cart.domain.Cart">
+    <update id="updateCart" parameterType="flab.commercemarket.domain.cart.vo.Cart">
         UPDATE cart
         SET
-            user_id = #{userId},
-            product_id = #{productId},
-            quantity = #{quantity}
+        user_id = #{userId},
+        product_id = #{productId},
+        quantity = #{quantity}
         WHERE
-            id = #{id}
+        id = #{id}
     </update>
 
     <select id="findById" parameterType="long" resultMap="cartResultMap">
